pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robot-mcp"
version = "0.3.0"
description = "Natural language robot control using FastMCP with multi-LLM support (OpenAI, Groq, Gemini, Ollama)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Daniel Gaida", email = "daniel.gaida@th-koeln.de"}
]
keywords = ["robotics", "mcp", "fastmcp", "llm", "groq", "openai", "gemini", "ollama", "pick-and-place", "natural-language"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
]

# Core dependencies - Multi-LLM support
dependencies = [
    "fastmcp>=0.1.0",
    "llm_client @ git+https://github.com/dgaida/llm_client.git",
    "python-dotenv>=1.0.0",
    "robot-environment @ git+https://github.com/dgaida/robot_environment.git",
    "text2speech @ git+https://github.com/dgaida/text2speech.git",
    "speech2text @ git+https://github.com/dgaida/speech2text.git",
]

[project.optional-dependencies]
# GUI dependencies
gui = [
    "gradio>=4.0.0",
    "redis_robot_comm @ git+https://github.com/dgaida/redis_robot_comm.git",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

# Documentation
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]

# All dependencies combined
all = [
    "robot-mcp[gui,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/dgaida/robot_mcp"
Documentation = "https://github.com/dgaida/robot_mcp/blob/main/README.md"
Repository = "https://github.com/dgaida/robot_mcp"
"Bug Tracker" = "https://github.com/dgaida/robot_mcp/issues"

[project.scripts]
# Command-line scripts
robot-fastmcp-server = "server.fastmcp_robot_server:main"
robot-universal-client = "client.fastmcp_universal_client:main"
robot-groq-client = "client.fastmcp_groq_client:main"
robot-gui = "robot_gui.mcp_app:main"
robot-examples = "examples.fastmcp_main_client:main"

[tool.setuptools]
packages = ["client", "server", "robot_gui", "examples"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.sh", "*.bat", "*.env.template"]

[tool.black]
line-length = 127
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 127
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.isort]
profile = "black"
line_length = 127
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["client", "server", "robot_gui"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
